+++ lmi_setup_02c.sh
@@ -44,8 +44,22 @@ findmnt -ro SOURCE,TARGET \
   | sed -e's,^[/A-Za-z0-9_-]*[[]\([^]]*\)[]],\1,' \
   | column -t
 
-rm --one-file-system --recursive --force \
-  "$(schroot --chroot=centos7lmi --location)"
+# Abort if any centos lmi mountpoint remains.
+#
+# This shouldn't occur, and 'rm --one-file-system' below should be
+# safe anyway, yet an actual catastrophe did occur nonetheless.
+findmnt | grep 'centos.*lmi' && exit 9
+
+loc0=/srv/chroot/centos7lmi
+loc1="$(schroot --chroot=centos7lmi --location)"
+if [ -n "${loc1}" ] && [ "${loc0}" != "${loc1}" ]; then
+  echo "chroot found in unexpected location--remove it manually"
+  exit 9
+fi
+
+rm --one-file-system --recursive --force "${loc0}"
+if [ -e "${loc0}" ] || [ -e "${loc1}" ] ; then echo "Oops."; exit 9; fi
+
 rm --force /etc/schroot/chroot.d/centos7lmi.conf
 
 stamp=$(date -u +'%Y%m%dT%H%M%SZ')
